Move CmdLineOverrides + parser from MRMcp to MRMcpSettings
Pure code motion — colocates CLI-flag parsing with the settings layer
(its only consumer). MRMcp public surface stays focused on the Server.
