+++ source/MRMcp/MRMcp.cpp
@@ -264,17 +264,4 @@ Server& getDefaultServer()
     return ret;
 }
 
-CmdLineOverrides parseCmdLineOverrides( const std::vector<std::string>& commandArgs )
-{
-    CmdLineOverrides out;
-    for ( size_t i = 0; i + 1 < commandArgs.size(); ++i )
-    {
-        if ( commandArgs[i] == "-mcpPort" )
-            out.port = std::atoi( commandArgs[i + 1].c_str() );
-        else if ( commandArgs[i] == "-mcpDumpFile" )
-            out.dumpFilePath = pathFromUtf8( commandArgs[i + 1] );
-    }
-    return out;
-}
-
 } // namespace MR
